backend/requirements.txt
# backend/requirements.txt
# Core Engine
pymupdf==1.23.26
sentence-transformers==2.7.0
torch>=2.3.0
nltk==3.8.1
huggingface-hub<0.21.0
scikit-learn # For entity extraction

# Backend Web Framework
fastapi
uvicorn[standard]
python-multipart

# Database
motor

# Settings Management
pydantic-settings

# External APIs
google-generativeai
azure-cognitiveservices-speech